Start using pyarrow instead of msgpack for serialization
We would previously rely on pyarrow for all things dataframe
serialization due to various bugs in the handling of datetimes in
pyarrow, but it looks like pyarrow is now finally stable enough that we
can start using it. As a result, we no longer have to pin the pandas
version.